+++ src/main/java/leets/leenk/domain/user/application/mapper/UserMapper.java
@@ -1,6 +1,5 @@
 package leets.leenk.domain.user.application.mapper;
 
-import leets.leenk.domain.user.application.dto.response.BirthdayUserResponse;
 import leets.leenk.domain.user.application.dto.response.UserInfoResponse;
 import leets.leenk.domain.user.domain.entity.User;
 import leets.leenk.global.auth.application.dto.response.OauthUserInfoResponse;
@@ -28,14 +27,6 @@ public class UserMapper {
                 .id(oauthUserInfoResponse.userId())
                 .name(oauthUserInfoResponse.name())
                 .cardinal(oauthUserInfoResponse.cardinal())
-                .build();
-    }
-
-    public BirthdayUserResponse toBirthdayUserResponse(User user) {
-        return BirthdayUserResponse.builder()
-                .id(user.getId())
-                .name(user.getName())
-                .profileImage(user.getProfileImage())
                 .build();
     }
 }
